successfully tested saving risks and reports to database
