+++ cybersecurity_assessment_tool/api/services/gemini_client.py
@@ -1,6 +1,6 @@
-import json
 from typing import Tuple, List, Optional
 from django.db import transaction
+from django.utils import timezone
 
 from ..models import Report, Risk, Organization, User
 from .ai_generation_service import ai_generation_service
@@ -24,6 +24,7 @@ from .ai_generation_service import ai_generation_service
 #         return JsonResponse({"status": "error", "message": "Failed to generate"}, status=500)
 
 # Define a sorting map for severities to translate strings to sortable integers
+
 SEVERITY_WEIGHTS = {
     "Critical": 1,
     "High": 2,
@@ -58,56 +59,113 @@ def build_current_risks_dict(organization_id: int) -> dict:
     }
     return current_risks
 
+def _inject_overview_and_questionnaire(report_data: dict, org: Organization) -> dict:
+    """
+    Injects the Overview and Questionnaire Review sections at the top 
+    of the AI-generated report data using information from the database.
+    """
+    new_section_data = {
+        "Overview": {
+            "Organization Name": org.org_name,
+            "Primary Domain": org.email_domain,
+            "External IP Address": org.external_ip,
+            "Report Date": timezone.now().strftime('%Y-%m-%d')
+        },
+        "Questionnaire Review": {
+            "Do you require MFA to access email?": "Yes" if org.require_mfa_email else "No",
+            "Do you require MFA to log into computers?": "Yes" if org.require_mfa_computer else "No",
+            "Do you require MFA to access sensitive data systems?": "Yes" if org.require_mfa_sensitive_data else "No",
+            "Does your organization have an employee acceptable use policy?": "Yes" if org.employee_acceptable_use_policy else "No",
+            "Does your organization do security awareness training for new employees?": "Yes" if org.training_new_employees else "No",
+            "Does your organization do security awareness training for all employees at least once per year?": "Yes" if org.training_once_per_year else "No"
+        }
+    }
+
+    if "report" in report_data and isinstance(report_data["report"], list):
+        for i, report_item in enumerate(report_data["report"]):
+            rebuilt_report_item = {}
+            
+            for key, value in new_section_data.items():
+                rebuilt_report_item[key] = value
+                
+            for key, value in report_item.items():
+                rebuilt_report_item[key] = value
+                
+            report_data["report"][i] = rebuilt_report_item
+
+    return report_data
+
 def generate_and_process_report(
-    organization_id: int, 
-    user_id: int, 
+    organization_id: str, 
+    user_id: str, 
     context_data: str
 ) -> Tuple[Optional[Report], Optional[List[Risk]]]:
     """
     Acts as the client to gather DB fields, call the AI service, 
-    sort the resulting risks by severity, and update the database.
+    sort the resulting data, inject database context, and save objects.
     """
     
-    # 1. Setup the inputs for the AI service
-    personal_info = {
-        "organization_id": organization_id,
-        "user_id": user_id
-    }
-    
+    # 1. Fetch current risks
     current_risks = build_current_risks_dict(organization_id)
 
-    # 2. Call the AI generation service
-    # Note: ai_generation_service creates the Report and Risk objects in the DB
-    report, created_risks = ai_generation_service(personal_info, current_risks, context_data)
+    # 2. Call the AI generation service (pure AI logic, no database IDs needed)
+    report_data, risks_data = ai_generation_service(current_risks, context_data)
 
-    if not report or not created_risks:
-        print("[ERROR] AI Service failed to generate report or risks.")
+    if report_data is None or risks_data is None:
+        print("[ERROR] AI Service failed to generate report or risks data.")
         return None, None
 
-    # 3. Sort the Python list of Risk objects (useful for returning to the frontend)
-    created_risks.sort(key=lambda r: get_severity_weight(r.severity))
-
-    # 4. Sort the JSON data inside the Report object
-    # We must navigate the schema: report -> [0] -> "Risks & Recommendations" -> "Vulnerabilities Found"
-    report_json = report.report_text
-    
+    # 3. Process and Save to Database
     try:
-        if "report" in report_json and isinstance(report_json["report"], list) and len(report_json["report"]) > 0:
-            readiness_section = report_json["report"][0].get("Risks & Recommendations", {})
-            vulnerabilities = readiness_section.get("Vulnerabilities Found", [])
-            
-            # Sort the actual JSON array in place
-            vulnerabilities.sort(key=lambda v: get_severity_weight(v.get("Severity", "")))
-            
-            # Re-assign back to ensure it's updated
-            report_json["report"][0]["Risks & Recommendations"]["Vulnerabilities Found"] = vulnerabilities
-            
-            # Update the report text and save the changes to the database
-            report.report_text = report_json
-            report.save(update_fields=['report_text'])
-            print(f"--- Successfully sorted vulnerabilities in Report {report.report_id} ---")
+        # Fetch the database records
+        org = Organization.objects.get(organization_id=organization_id)
+        user = User.objects.get(user_id=user_id) if user_id else None
 
+        # Inject context from the database into the AI's output
+        report_data = _inject_overview_and_questionnaire(report_data, org)
+
+        with transaction.atomic():
+            # Sort the JSON vulnerabilities BEFORE saving to the database
+            if "report" in report_data and isinstance(report_data["report"], list) and len(report_data["report"]) > 0:
+                readiness_section = report_data["report"][0].get("Risks & Recommendations", {})
+                vulnerabilities = readiness_section.get("Vulnerabilities Found", [])
+                
+                vulnerabilities.sort(key=lambda v: get_severity_weight(v.get("Severity", "")))
+                report_data["report"][0]["Risks & Recommendations"]["Vulnerabilities Found"] = vulnerabilities
+
+            # Create the Report
+            new_report = Report.objects.create(
+                user_created=user,
+                organization=org,
+                report_name=f"Report - {org.org_name} - {timezone.now().strftime('%Y-%m-%d')}",
+                report_text=report_data, 
+                completed=timezone.now()
+            )
+
+            # Create the Risks
+            created_risks = []
+            for risk_item in risks_data.get('new vulnerabilities', []):
+                new_risk = Risk.objects.create(
+                    risk_name=risk_item.get('risk_name'),
+                    report=new_report, 
+                    organization=org,
+                    overview=risk_item.get('overview'),
+                    recommendations=risk_item.get('recommendations'),
+                    severity=risk_item.get('severity'),
+                    affected_elements=", ".join(risk_item.get('affected_elements', [])),
+                )
+                created_risks.append(new_risk)
+
+        print(f"--- Successfully saved Report {new_report.pk} and associated risks. ---")
+
+        # 4. Sort the Python list of Risk objects for returning to the frontend
+        created_risks.sort(key=lambda r: get_severity_weight(r.severity))
+
+        return new_report, created_risks
+
+    except Organization.DoesNotExist:
+        print(f"[ERROR] Organization with ID {organization_id} not found.")
+        return None, None
     except Exception as e:
-        print(f"[WARNING] Could not sort JSON vulnerabilities: {e}")
-
-    return report, created_risks
+        print(f"[ERROR] Database save failed: {e}")
+        return None, None
