+++ cybersecurity_assessment_tool/scripts/test_gemini.py
@@ -1,5 +1,6 @@
 import json
-from api.services.ai_generation_service import ai_generation_service
+# Update this import path if gemini_client is located elsewhere in your app structure
+from api.services.gemini_client import generate_and_process_report
 
 def run(*args):
     """
@@ -11,38 +12,21 @@ def run(*args):
     # 1. Context is the raw technical data/prompt you want Gemini to analyze.
     # We grab it from args if provided, otherwise use a default test string.
     context = args[0] if args else (
-        "Assess a company that has no MFA, reuses admin passwords, "
-        "has port 3389 open to the internet, and lacks DMARC records."
+        # "Assess a company that has no MFA, reuses admin passwords,"
+        # "does not have the port 3389 open to the internet, and does not lack DMARC records."
+        "Assess a company that is running an outdated, highly vulnerable FTP server on port 21 "
+        "and uses default administrator credentials for their local router."
     )
     
-    # 2. personal_info needs to be a dict with IDs that exist in your local database!
-    # UPDATE THESE IDs to match a valid Organization and User in your DB.
-    personal_info = {
-        "organization_id": "68e1d320-1fe3-4f14-8d1f-760e9dc1da82", 
-        "user_id": "58f961f8-a5e9-4e58-9efb-8b066abd9f28"
-    }
+    # 2. UPDATE THESE IDs to match a valid Organization and User in your DB.
+    # The client will use these to build the personal_info dict and fetch current_risks.
+    organization_id = "68e1d320-1fe3-4f14-8d1f-760e9dc1da82" 
+    user_id = "58f961f8-a5e9-4e58-9efb-8b066abd9f28"
     
-    # 3. current_risks simulates the existing risks the org already knows about.
-    # This gives Gemini something to cross-reference against to avoid duplicates.
-    current_risks = {
-        "known_vulnerabilities": [
-            {
-                "risk_name": "Lack of DMARC",
-                "overview": "Domain is missing DMARC record, allowing email spoofing.",
-                "severity": "Medium",
-                "affected_elements": ["email domain"],
-                "recommendations": {
-                    "easy_fix": "Add a basic p=none DMARC record.",
-                    "long_term_fix": "Enforce p=reject."
-                }
-            }
-        ]
-    }
+    print(f"\n--- Calling Gemini Client with context ---\n{context}\n")
 
-    print(f"\n--- Calling Gemini with context ---\n{context}\n")
-
-    # ai_generation_service returns a tuple: (Report object, list of Risk objects)
-    report, risks = ai_generation_service(personal_info, current_risks, context)
+    # 3. generate_and_process_report returns a tuple: (Report object, sorted list of Risk objects)
+    report, risks = generate_and_process_report(organization_id, user_id, context)
 
     print("\n--- Generation Complete ---\n")
     
@@ -54,7 +38,7 @@ def run(*args):
         print("❌ Report generation failed or returned None.")
 
     if risks:
-        print(f"✅ Created {len(risks)} new risks:")
+        print(f"✅ Created {len(risks)} new risks (sorted by severity):")
         for r in risks:
             print(f"  - [{r.severity}] {r.risk_name}")
     else:
